+++ src/main/java/net/silentchaos512/loginar/block/urn/LoginarUrnSwapperMenu.java
@@ -12,7 +12,7 @@ import net.neoforged.neoforge.items.IItemHandler;
 import net.neoforged.neoforge.items.ItemHandlerCopySlot;
 import net.neoforged.neoforge.network.PacketDistributor;
 import net.silentchaos512.loginar.LoginarMod;
-import net.silentchaos512.loginar.network.CPacketSwapItemFromUrn;
+import net.silentchaos512.loginar.network.SwapItemFromUrnPayload;
 import net.silentchaos512.loginar.setup.LsDataComponents;
 import net.silentchaos512.loginar.setup.LsMenuTypes;
 
@@ -52,7 +52,7 @@ public class LoginarUrnSwapperMenu extends AbstractContainerMenu {
             ItemStack item = slot.getItem();
 
             LoginarMod.LOGGER.info("Attempting to swap urn item with hand: {}", item);
-            PacketDistributor.sendToServer(new CPacketSwapItemFromUrn(slotIndex));
+            PacketDistributor.sendToServer(new SwapItemFromUrnPayload(slotIndex));
             player.closeContainer();
         }
     }