+++ src/main/java/net/silentchaos512/loginar/network/SwapItemFromUrnPayload.java
@@ -6,12 +6,12 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.silentchaos512.loginar.LoginarMod;
 
-public record CPacketSwapItemFromUrn(int urnItemSlot) implements CustomPacketPayload {
-    public static final Type<CPacketSwapItemFromUrn> TYPE = new Type<>(LoginarMod.getId("swap_item_from_urn"));
+public record SwapItemFromUrnPayload(int urnItemSlot) implements CustomPacketPayload {
+    public static final Type<SwapItemFromUrnPayload> TYPE = new Type<>(LoginarMod.getId("swap_item_from_urn"));
 
-    public static final StreamCodec<FriendlyByteBuf, CPacketSwapItemFromUrn> STREAM_CODEC = StreamCodec.composite(
+    public static final StreamCodec<FriendlyByteBuf, SwapItemFromUrnPayload> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.VAR_INT, d -> d.urnItemSlot,
-            CPacketSwapItemFromUrn::new
+            SwapItemFromUrnPayload::new
     );
 
     @Override